[Week1] 코드리뷰 반영
Extension 변경
